requirements.txt
google-cloud-bigquery>=3.11.0
pandas>=2.0.0
numpy>=1.24.0
db-dtypes>=1.1.0
torch>=2.1.0
pytorch-tabnet>=4.1.0
scikit-learn>=1.3.0
shap>=0.44.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas-gbq>=0.19.0
ipykernel>=6.0.0